+++ inc/frontend/shortcodes/bandsintown.php
@@ -23,6 +23,9 @@ if ( ! function_exists( 'wpb_shortcode_bandsintown_events' ) ) {
 
 		extract( shortcode_atts( array(
 			'artist' => '',
+			'local_dates' => 'true',
+			'past_dates' => 'true',
+			'display_limit' => '',
 			'inline_style' => '',
 			'extra_class' => '',
 		), $atts ) );
@@ -31,21 +34,37 @@ if ( ! function_exists( 'wpb_shortcode_bandsintown_events' ) ) {
 		$artist = sanitize_text_field( $artist );
 		$artist_slug = sanitize_title( $artist );
 
+		//wp_enqueue_script( 'bandsintown', 'https://widget.bandsintown.com/main.min.js', array(), false, true );
+
 		if ( $artist ) {
 
-			ob_start(); ?>
-			<script type="text/javascript" src="https://widget.bandsintown.com/javascripts/bit_widget.js"></script>
-			<script type="text/javascript">  
-			jQuery( document ).ready( function( $ ) { 
-				new BIT.Widget( { 
-					'artist' : '<?php echo esc_js( $artist ); ?>',
-					'div_id' : 'wpb-bandwintown-tour-dates-<?php echo sanitize_html_class( $artist_slug ); ?>'
-				} ).insert_events();
-			} );
-			</script>
-			<div id="wpb-bandwintown-tour-dates-<?php echo sanitize_html_class( $artist_slug ); ?>"></div>
-			<?php
-			$output = ob_get_clean();
+			$accent_color = ( function_exists( 'wolf_get_theme_mod' ) ) ? wolf_get_theme_mod( 'accent_color', '#0073AA' ) : '#0073AA';
+
+			$options = array(
+				'artist' => $artist,
+				'text_color' => '',
+				'background_color' => '',
+				'display_limit' => $display_limit,
+				'link_text_color' => '#ffffff',
+				'link_color' => $accent_color,
+				'local_dates' => $local_dates,
+				'past_dates' => $past_dates,
+			);
+
+			$output .= '<script type="text/javascript" src="https://widget.bandsintown.com/main.min.js"></script>';
+			$output .= '<a class="bit-widget-initializer"'
+			. 'data-text-color="' . esc_attr( $options['text_color'] ) . '"'
+			. 'data-background-color="' . esc_attr( $options['background_color'] ) . '"'
+			. 'data-display-limit="' . esc_attr( $options['display_limit'] ) . '"'
+			. 'data-link-text-color="' . esc_attr( $options['link_text_color' ] ) . '"'
+			. 'data-popup-background-color="#FFFFFF"'
+			. 'data-artist-name="' . esc_attr( $options['artist'] ) . '"'
+			. 'data-link-color="' . esc_attr( $options['link_color'] ) . '"'
+			. 'data-display-local-dates="' . esc_attr( $options['local_dates'] ) . '"'
+			. 'data-display-past-dates="' . esc_attr( $options['past_dates'] ) . '"'
+			. 'data-auto-style="false"'
+			;
+			$output .= '></a>';
 
 		} else {
 			if ( is_user_logged_in() ) {